% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.iv_test}
\alias{format.iv_test}
\title{Format method for an IV-validity test}
\usage{
\method{format}{iv_test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{iv_test}.}

\item{...}{Ignored.}
}
\value{
A one-line character summary.
}
\description{
Used when an \code{iv_test} object is included as a column of a data frame
or tibble.
}
