% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.iv_test}
\alias{plot.iv_test}
\title{Plot method for an IV-validity test}
\usage{
\method{plot}{iv_test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{iv_test}.}

\item{...}{Further graphical arguments passed to \link[graphics:hist]{graphics::hist}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plots the bootstrap distribution of the test statistic with the
observed statistic and the rejection region highlighted.
}
