% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.iv_check}
\alias{print.iv_check}
\title{Print method for an iv_check result}
\usage{
\method{print}{iv_check}(x, digits = 3L, ...)
}
\arguments{
\item{x}{An object of class \code{iv_check}.}

\item{digits}{Number of significant digits.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for an iv_check result
}
