% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.iv_test}
\alias{print.iv_test}
\title{Print method for an IV-validity test}
\usage{
\method{print}{iv_test}(x, digits = 3L, ...)
}
\arguments{
\item{x}{An object of class \code{iv_test}.}

\item{digits}{Number of significant digits to display.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for an IV-validity test
}
