% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.iv_test}
\alias{summary.iv_test}
\title{Summary method for an IV-validity test}
\usage{
\method{summary}{iv_test}(object, ...)
}
\arguments{
\item{object}{An object of class \code{iv_test}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for an IV-validity test
}
